* Killed a leak in the theora encoder. * Added a repaint function to dsplay * Killed a few warnings * Removed an ignore warning pragma which actually generated even more warnings for some reason. * Token tidy up and documentation in OOTheora.
git-svn-id: https://svn.xiph.org/trunk/oggdsf@8675 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/core/directshow/dsfOggDemux/HTTPFileSource.cpp
@@ -120,7 +120,7 @@ void HTTPFileSource::unChunk(unsigned char* inBuff, unsigned long inNumBytes)
 
 				debugLog<<"Chunk reamining "<<mChunkRemains<<endl;
 				//The size of the crlf 's and the chunk size value
-				unsigned long locGuffSize = (locChunkSizeStr.size() + 2);
+				unsigned long locGuffSize = (unsigned long)(locChunkSizeStr.size() + 2);
 				locWorkingBuffPtr +=  locGuffSize;
 				locNumBytesLeft -= locGuffSize;
 			} else {
